upload & download files

+++ controllers/users.js
@@ -1,9 +1,13 @@
 const bcrypt = require('bcryptjs');
 var nodemailer = require('nodemailer');
 const { validationResult } = require('express-validator');
+const fs = require('fs');
+const path = require('path');
+const PDFDocument = require('pdfkit');
 
 const User = require('../models/user');
 const Role = require('../models/role');
+const fileHelper = require('../util/file');
 
 var transporter = nodemailer.createTransport({
     service: 'gmail',
@@ -67,6 +71,7 @@ exports.postAddUser = (req, res, next) => {
     const password = req.body.password;
     const description = req.body.description;
     const roleId = req.body.roleId;
+    const image = req.file;
 
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
@@ -79,6 +84,7 @@ exports.postAddUser = (req, res, next) => {
                 validationErrors: errors.array()
             });
         }).catch(err => {
+            fileHelper.deleteFile(req.file?.path);
             const error = new Error(err);
             error.httpStatusCode = 500;
             return next(error);
@@ -92,7 +98,8 @@ exports.postAddUser = (req, res, next) => {
                     password: hashedPassword,
                     age: age,
                     description: description,
-                    roleId: roleId
+                    roleId: roleId,
+                    imageUrl: image?.path
                 });
             })
             .then((user) => {
@@ -107,15 +114,15 @@ exports.postAddUser = (req, res, next) => {
                 <p><b>Age: </b>${user.age}</p>`
                 });
             })
-            .then((error, info) => {
-                if (error) {
-                    const error = new Error(err);
-                    error.httpStatusCode = 500;
-                    return next(error);
-                } else {
-                    console.info('Email sent: ' + info.response);
-                }
-            })
+            // .then((error, info) => {
+            //     if (error) {
+            //         const error = new Error(err);
+            //         error.httpStatusCode = 500;
+            //         return next(error);
+            //     } else {
+            //         console.info('Email sent: ' + info.response);
+            //     }
+            // })
             .catch(err => {
                 const error = new Error(err);
                 error.httpStatusCode = 500;
@@ -171,17 +178,27 @@ exports.postEditUser = (req, res, next) => {
     else {
         bcrypt.hash(req.body.password, 12)
             .then(hashedPassword => {
+                let imageUrl;
+                if (req.file) {
+                    fileHelper.deleteFile(req.body.imageUrl);
+                    imageUrl = req.file.path;
+                }
+                else {
+                    imageUrl = req.body.imageUrl;
+                }
                 return User.updateOne({ _id: req.body.id },
                     {
                         name: req.body.name,
                         password: hashedPassword,
                         age: req.body.age,
                         description: req.body.description,
-                        roleId: req.body.roleId
+                        roleId: req.body.roleId,
+                        imageUrl: imageUrl
                     });
             })
             .then(() => res.redirect('/users-list'))
             .catch(err => {
+                fileHelper.deleteFile(req.file.path);
                 const error = new Error(err);
                 error.httpStatusCode = 500;
                 return next(error);
@@ -190,11 +207,63 @@ exports.postEditUser = (req, res, next) => {
 };
 
 exports.postDeleteUser = (req, res, next) => {
-    User.deleteOne({ _id: req.params.userId })
+    User.findById(req.params.userId)
+        .then(user => {
+            if (!user) {
+                return next(new Error('User not found.'));
+            }
+            fileHelper.deleteFile(user.imageUrl);
+            return User.deleteOne({ _id: req.params.userId });
+        })
         .then(() => res.redirect('/users-list'))
         .catch(err => {
             const error = new Error(err);
             error.httpStatusCode = 500;
             return next(error);
         });
+};
+
+exports.getPdfUser = (req, res, next) => {
+    const userId = req.params.userId;
+    User.findById(userId)
+        .populate('roleId', 'name')
+        .then(user => {
+            if (!user) {
+                return next(new Error('No user found.'));
+            }
+            const userFileName = 'user-' + userId + '.pdf';
+            const userFilePath = path.join('data', 'pdf', userFileName);
+
+            const pdfDoc = new PDFDocument();
+            res.setHeader('Content-Type', 'application/pdf');
+            res.setHeader(
+                'Content-Disposition',
+                'inline; filename="' + userFileName + '"'
+            );
+            pdfDoc.pipe(fs.createWriteStream(userFilePath));
+            pdfDoc.pipe(res);
+
+            pdfDoc.fontSize(30).text(user.name, {
+                underline: true
+            });
+            pdfDoc.text('-----------------------');
+            pdfDoc.font('Helvetica-Bold').text('Age: ', {
+                continued: true
+            }).font('Helvetica').text(user.age);
+            pdfDoc.font('Helvetica-Bold').text('Description: ', {
+                continued: true
+            }).font('Helvetica').text(user.description);
+            pdfDoc.font('Helvetica-Bold').text('Role: ', {
+                continued: true
+            }).font('Helvetica').text(user.roleId.name);
+
+            pdfDoc.text('---');
+            if (user.imageUrl)
+                pdfDoc.image(user.imageUrl, 430, 15, { fit: [100, 100], align: 'center', valign: 'center' })
+                    .rect(430, 15, 100, 100).stroke()
+                    .text('', 430, 0);
+
+            pdfDoc.end();
+        })
+        .catch(err => next(err));
 };
